Add history menu with basic details Include history for current year to report

+++ src/main/java/cz/kostka/rybyjstr/service/CatchService.java
@@ -154,7 +154,7 @@ public class CatchService {
         return catchRepository.count();
     }
 
-    public long getAllCatchesSize() {
+    public int getAllCatchesSize() {
         return catchRepository.sumAllSizes();
     }
